+++ src/components/new-safe/create/index.tsx
@@ -2,21 +2,21 @@ import { Container, Typography, Grid } from '@mui/material'
 import { useRouter } from 'next/router'
 
 import useWallet from '@/hooks/wallets/useWallet'
-import OverviewWidget from '../OverviewWidget'
+import OverviewWidget from '@/components/new-safe/create/OverviewWidget'
 import type { NamedAddress } from '@/components/create-safe/types'
-import type { TxStepperProps } from '../CardStepper/useCardStepper'
-import CreateSafeStep0 from '@/components/new-safe/steps/Step0'
-import CreateSafeStep1 from '@/components/new-safe/steps/Step1'
-import CreateSafeStep2 from '@/components/new-safe/steps/Step2'
-import CreateSafeStep3 from '@/components/new-safe/steps/Step3'
-import { CreateSafeStatus } from '@/components/new-safe/steps/Step4'
+import type { TxStepperProps } from '@/components/new-safe/CardStepper/useCardStepper'
+import CreateSafeStep0 from '@/components/new-safe/create/steps/Step0'
+import CreateSafeStep1 from '@/components/new-safe/create/steps/Step1'
+import CreateSafeStep2 from '@/components/new-safe/create/steps/Step2'
+import CreateSafeStep3 from '@/components/new-safe/create/steps/Step3'
+import { CreateSafeStatus } from '@/components/new-safe/create/steps/Step4'
 import useAddressBook from '@/hooks/useAddressBook'
-import { CardStepper } from '../CardStepper'
+import { CardStepper } from '@/components/new-safe/CardStepper'
 import { AppRoutes } from '@/config/routes'
 import { CREATE_SAFE_CATEGORY } from '@/services/analytics'
 import type { AlertColor } from '@mui/material'
-import type { CreateSafeInfoItem } from '../CreateSafeInfos'
-import CreateSafeInfos from '../CreateSafeInfos'
+import type { CreateSafeInfoItem } from '@/components/new-safe/create/CreateSafeInfos'
+import CreateSafeInfos from '@/components/new-safe/create/CreateSafeInfos'
 import { type ReactElement, useMemo, useState } from 'react'
 import ExternalLink from '@/components/common/ExternalLink'
