+++ src/components/new-safe/create/steps/Step0/index.tsx
@@ -4,12 +4,12 @@ import useWallet from '@/hooks/wallets/useWallet'
 import { useCurrentChain } from '@/hooks/useChains'
 import { isPairingSupported } from '@/services/pairing/utils'
 
-import type { NewSafeFormData } from '@/components/new-safe/CreateSafe'
+import type { NewSafeFormData } from '@/components/new-safe/create'
 import type { StepRenderProps } from '@/components/new-safe/CardStepper/useCardStepper'
-import useSyncSafeCreationStep from '@/components/new-safe/CreateSafe/useSyncSafeCreationStep'
-import layoutCss from '@/components/new-safe/CreateSafe/styles.module.css'
+import useSyncSafeCreationStep from '@/components/new-safe/create/useSyncSafeCreationStep'
+import layoutCss from '@/components/new-safe/create/styles.module.css'
 import useLocalStorage from '@/services/local-storage/useLocalStorage'
-import { type PendingSafeData, SAFE_PENDING_CREATION_STORAGE_KEY } from '@/components/new-safe/steps/Step4'
+import { type PendingSafeData, SAFE_PENDING_CREATION_STORAGE_KEY } from '@/components/new-safe/create/steps/Step4'
 import useConnectWallet from '@/components/common/ConnectWallet/useConnectWallet'
 import KeyholeIcon from '@/components/common/icons/KeyholeIcon'
 import PairingDescription from '@/components/common/PairingDetails/PairingDescription'
