+++ src/components/new-safe/create/steps/Step1/index.tsx
@@ -3,14 +3,14 @@ import { FormProvider, useForm } from 'react-hook-form'
 import { useMnemonicSafeName } from '@/hooks/useMnemonicName'
 import InfoIcon from '@/public/images/notifications/info.svg'
 import NetworkSelector from '@/components/common/NetworkSelector'
-import type { StepRenderProps } from '../../CardStepper/useCardStepper'
-import type { NewSafeFormData } from '../../CreateSafe'
-import useSyncSafeCreationStep from '@/components/new-safe/CreateSafe/useSyncSafeCreationStep'
+import type { StepRenderProps } from 'src/components/new-safe/CardStepper/useCardStepper'
+import type { NewSafeFormData } from '@/components/new-safe/create'
+import useSyncSafeCreationStep from '@/components/new-safe/create/useSyncSafeCreationStep'
 
-import css from './styles.module.css'
-import layoutCss from '@/components/new-safe/CreateSafe/styles.module.css'
+import css from 'src/components/new-safe/create/steps/Step1/styles.module.css'
+import layoutCss from '@/components/new-safe/create/styles.module.css'
 import useIsWrongChain from '@/hooks/useIsWrongChain'
-import NetworkWarning from '@/components/new-safe/NetworkWarning'
+import NetworkWarning from '@/components/new-safe/create/NetworkWarning'
 import NameInput from '@/components/common/NameInput'
 import { CREATE_SAFE_EVENTS, trackEvent } from '@/services/analytics'
 import ExternalLink from '@/components/common/ExternalLink'
