+++ src/components/new-safe/create/steps/Step2/index.tsx
@@ -4,17 +4,17 @@ import type { ReactElement } from 'react'
 
 import AddIcon from '@/public/images/common/add.svg'
 import InfoIcon from '@/public/images/notifications/info.svg'
-import { OwnerRow } from './OwnerRow'
+import { OwnerRow } from 'src/components/new-safe/create/steps/Step2/OwnerRow'
 import type { NamedAddress } from '@/components/create-safe/types'
-import type { StepRenderProps } from '../../CardStepper/useCardStepper'
-import type { NewSafeFormData } from '../../CreateSafe'
-import type { CreateSafeInfoItem } from '../../CreateSafeInfos'
-import { useSafeSetupHints } from './useSafeSetupHints'
-import useSyncSafeCreationStep from '@/components/new-safe/CreateSafe/useSyncSafeCreationStep'
+import type { StepRenderProps } from 'src/components/new-safe/CardStepper/useCardStepper'
+import type { NewSafeFormData } from '@/components/new-safe/create'
+import type { CreateSafeInfoItem } from 'src/components/new-safe/create/CreateSafeInfos'
+import { useSafeSetupHints } from 'src/components/new-safe/create/steps/Step2/useSafeSetupHints'
+import useSyncSafeCreationStep from '@/components/new-safe/create/useSyncSafeCreationStep'
 import ArrowBackIcon from '@mui/icons-material/ArrowBack'
-import css from './styles.module.css'
-import layoutCss from '@/components/new-safe/CreateSafe/styles.module.css'
-import NetworkWarning from '@/components/new-safe/NetworkWarning'
+import css from 'src/components/new-safe/create/steps/Step2/styles.module.css'
+import layoutCss from '@/components/new-safe/create/styles.module.css'
+import NetworkWarning from '@/components/new-safe/create/NetworkWarning'
 import useIsWrongChain from '@/hooks/useIsWrongChain'
 import { CREATE_SAFE_EVENTS, trackEvent } from '@/services/analytics'
