+++ src/components/new-safe/create/steps/Step2/styles.module.css
@@ -10,3 +10,15 @@
 .select :global .MuiSelect-select {
   padding: 12px 14px;
 }
+
+.name :global .MuiFormHelperText-root {
+  position: absolute;
+  bottom: -20px;
+}
+
+@media (max-width: 900px) {
+  .name :global .MuiFormHelperText-root {
+    position: relative;
+    bottom: 0;
+  }
+}
