+++ src/components/new-safe/create/steps/Step3/index.tsx
@@ -7,18 +7,18 @@ import useGasPrice from '@/hooks/useGasPrice'
 import { useEstimateSafeCreationGas } from '@/components/create-safe/useEstimateSafeCreationGas'
 import { formatVisualAmount } from '@/utils/formatters'
 import type { StepRenderProps } from '@/components/new-safe/CardStepper/useCardStepper'
-import type { NewSafeFormData } from '@/components/new-safe/CreateSafe'
-import css from './styles.module.css'
-import layoutCss from '@/components/new-safe/CreateSafe/styles.module.css'
+import type { NewSafeFormData } from '@/components/new-safe/create'
+import css from 'src/components/new-safe/create/steps/Step3/styles.module.css'
+import layoutCss from '@/components/new-safe/create/styles.module.css'
 import { getFallbackHandlerContractInstance } from '@/services/contracts/safeContracts'
 import { computeNewSafeAddress } from '@/components/create-safe/logic'
 import useWallet from '@/hooks/wallets/useWallet'
 import { useWeb3 } from '@/hooks/wallets/web3'
 import useLocalStorage from '@/services/local-storage/useLocalStorage'
-import { type PendingSafeData, SAFE_PENDING_CREATION_STORAGE_KEY } from '@/components/new-safe/steps/Step4'
-import useSyncSafeCreationStep from '@/components/new-safe/CreateSafe/useSyncSafeCreationStep'
+import { type PendingSafeData, SAFE_PENDING_CREATION_STORAGE_KEY } from '@/components/new-safe/create/steps/Step4'
+import useSyncSafeCreationStep from '@/components/new-safe/create/useSyncSafeCreationStep'
 import ArrowBackIcon from '@mui/icons-material/ArrowBack'
-import NetworkWarning from '@/components/new-safe/NetworkWarning'
+import NetworkWarning from '@/components/new-safe/create/NetworkWarning'
 import useIsWrongChain from '@/hooks/useIsWrongChain'
 import palette from '@/styles/colors'
