+++ src/components/new-safe/create/steps/Step3/styles.module.css
@@ -2,4 +2,5 @@
   display: flex;
   flex-direction: column;
   gap: var(--space-2);
+  font-size: 14px;
 }
