+++ src/components/new-safe/create/steps/Step4/LoadingSpinner/index.tsx
@@ -1,7 +1,7 @@
 import { Box } from '@mui/material'
-import css from './styles.module.css'
+import css from 'src/components/new-safe/create/steps/Step4/LoadingSpinner/styles.module.css'
 import classnames from 'classnames'
-import { SafeCreationStatus } from '@/components/new-safe/steps/Step4/useSafeCreation'
+import { SafeCreationStatus } from '@/components/new-safe/create/steps/Step4/useSafeCreation'
 import { useCallback, useEffect, useRef } from 'react'
 
 const rectTlEndTransform = 'translateX(0) translateY(20px) scaleY(1.1)'
