+++ src/components/new-safe/create/steps/Step4/index.tsx
@@ -5,19 +5,19 @@ import { useRouter } from 'next/router'
 import Track from '@/components/common/Track'
 import { CREATE_SAFE_EVENTS } from '@/services/analytics/events/createLoadSafe'
 import useLocalStorage from '@/services/local-storage/useLocalStorage'
-import StatusMessage from '@/components/new-safe/steps/Step4/StatusMessage'
+import StatusMessage from '@/components/new-safe/create/steps/Step4/StatusMessage'
 import useWallet from '@/hooks/wallets/useWallet'
 import useIsWrongChain from '@/hooks/useIsWrongChain'
-import type { NewSafeFormData } from '@/components/new-safe/CreateSafe'
+import type { NewSafeFormData } from '@/components/new-safe/create'
 import type { StepRenderProps } from '@/components/new-safe/CardStepper/useCardStepper'
-import type { PendingSafeTx } from '@/components/create-safe/types.d'
-import useSafeCreationEffects from '@/components/new-safe/steps/Step4/useSafeCreationEffects'
-import { SafeCreationStatus, useSafeCreation } from '@/components/new-safe/steps/Step4/useSafeCreation'
-import StatusStepper from '@/components/new-safe/steps/Step4/StatusStepper'
+import type { PendingSafeTx } from '@/components/create-safe/types'
+import useSafeCreationEffects from '@/components/new-safe/create/steps/Step4/useSafeCreationEffects'
+import { SafeCreationStatus, useSafeCreation } from '@/components/new-safe/create/steps/Step4/useSafeCreation'
+import StatusStepper from '@/components/new-safe/create/steps/Step4/StatusStepper'
 import { trackEvent } from '@/services/analytics'
 import useChainId from '@/hooks/useChainId'
-import { getRedirect } from '@/components/new-safe/steps/Step4/logic'
-import layoutCss from '@/components/new-safe/CreateSafe/styles.module.css'
+import { getRedirect } from '@/components/new-safe/create/steps/Step4/logic'
+import layoutCss from '@/components/new-safe/create/styles.module.css'
 import { AppRoutes } from '@/config/routes'
 import palette from '@/styles/colors'
