+++ src/components/new-safe/create/steps/Step4/logic/index.ts
@@ -3,13 +3,13 @@ import type { ChainInfo } from '@safe-global/safe-gateway-typescript-sdk'
 import { getProxyFactoryContractInstance } from '@/services/contracts/safeContracts'
 import type { ConnectedWallet } from '@/services/onboard'
 import { BigNumber } from '@ethersproject/bignumber'
-import { SafeCreationStatus } from '@/components/new-safe/steps/Step4/useSafeCreation'
+import { SafeCreationStatus } from '@/components/new-safe/create/steps/Step4/useSafeCreation'
 import { didRevert, type EthersError } from '@/utils/ethers-utils'
 import { Errors, logError } from '@/services/exceptions'
 import { ErrorCode } from '@ethersproject/logger'
 import { isWalletRejection } from '@/utils/wallets'
 import type { PendingSafeTx } from '@/components/create-safe/types'
-import type { NewSafeFormData } from '@/components/new-safe/CreateSafe'
+import type { NewSafeFormData } from '@/components/new-safe/create'
 import type { UrlObject } from 'url'
 import chains from '@/config/chains'
 import { AppRoutes } from '@/config/routes'
