+++ src/components/new-safe/create/steps/Step4/useSafeCreation.ts
@@ -4,7 +4,7 @@ import { useWeb3, useWeb3ReadOnly } from '@/hooks/wallets/web3'
 import { useCurrentChain } from '@/hooks/useChains'
 import useWallet from '@/hooks/wallets/useWallet'
 import type { EthersError } from '@/utils/ethers-utils'
-import type { PendingSafeData } from '@/components/new-safe/steps/Step4/index'
+import type { PendingSafeData } from '@/components/new-safe/create/steps/Step4/index'
 import type { PendingSafeTx } from '@/components/create-safe/types'
 import {
   checkSafeCreationTx,
@@ -12,7 +12,7 @@ import {
   handleSafeCreationError,
   SAFE_CREATION_ERROR_KEY,
   showSafeCreationError,
-} from '@/components/new-safe/steps/Step4/logic'
+} from '@/components/new-safe/create/steps/Step4/logic'
 import { useAppDispatch } from '@/store'
 import { closeByGroupKey } from '@/store/notificationsSlice'
 import { CREATE_SAFE_EVENTS, trackEvent } from '@/services/analytics'
